requirements.txt
flask>=3.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
python-escpos>=3.0
qrcode[pil]>=7.4
pillow>=10.0
